// src/app/features/auth/pages/register/register.page.ts
import { Component, inject } from '@angular/core';
import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
import { ButtonModule } from 'primeng/button';
import { CardModule } from 'primeng/card';
import { InputTextModule } from 'primeng/inputtext';
import { PasswordModule } from 'primeng/password';
import { AuthStore } from '../../../../core/auth/auth.store';
import { RegisterForm } from '../../models/auth-model';
import { passwordMatchValidator } from '../../../../shared/validators/password-match-validator';
import { RouterLink } from '@angular/router';
import { first } from 'rxjs';
@Component({
  selector: 'app-register',
  standalone: true,
  imports: [
    ReactiveFormsModule,
    ButtonModule,
    InputTextModule,
    PasswordModule,
    CardModule,
    RouterLink,
  ],
  templateUrl: './register.page.html',
  styleUrl: './register.page.scss',
})
export class RegisterPage {
  private readonly fb = inject(FormBuilder);
  protected readonly authStore = inject(AuthStore);

  registerForm: FormGroup<RegisterForm> = this.fb.nonNullable.group(
    {
      firstName: this.fb.nonNullable.control('', [Validators.required, Validators.maxLength(50)]),
      lastName: this.fb.nonNullable.control('', [Validators.required, Validators.maxLength(50)]),
      email: this.fb.nonNullable.control('', [
        Validators.required,
        Validators.email,
        Validators.maxLength(150),
      ]),
      password: this.fb.nonNullable.control('', [
        Validators.required,
        Validators.minLength(8),
        Validators.maxLength(50),
      ]),
      repeatPassword: this.fb.nonNullable.control('', [
        Validators.required,
        Validators.minLength(8),
        Validators.maxLength(50),
      ]),
    },
    { validators: passwordMatchValidator('password', 'repeatPassword') },
  );

  onSubmit(): void {
    this.registerForm.markAllAsTouched();
    if (this.registerForm.invalid) return;

    const formData = this.registerForm.getRawValue();

    const registerData = {
      firstName: formData.firstName,
      lastName: formData.lastName,
      email: formData.email,
      password: formData.password,
    };
    
    this.authStore.register(registerData);
  }

  protected get firstName() {
    return this.registerForm.controls.firstName;
  }
  protected get lastName() {
    return this.registerForm.controls.lastName;
  }
  protected get email() {
    return this.registerForm.controls.email;
  }
  protected get password() {
    return this.registerForm.controls.password;
  }
  protected get repeatPassword() {
    return this.registerForm.controls.repeatPassword;
  }
}
